pyproject.toml
[tool.poetry]
name = "realtime-ai-dev"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
gradio = "^4.37.2"
pandas = "^2.2.2"
aiohttp = "^3.9.5"
dataclasses-json = "^0.6.7"
requests = "^2.32.3"
pydantic = "^2.8.2"
python-dotenv = "^1.0.1"
sseclient-py = "^1.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
